fix: group sprite system on minimum components only
This commit fixes an issue where the sprite system forces an entity to have a size and color component in order to be rendered. The size and color components are optional. They are unnecessary for an entity with a sprite to be successfully rendered. This commit also makes the sprite system provide default size and color components.